requirements.txt
contextlib2==21.6.0
marshmallow==3.20.2
numpy==1.26.4
packaging==23.2
paho-mqtt==1.6.1
pandas
python-dateutil==2.8.2
pytz==2024.1
schema==0.7.5
six==1.16.0
tzdata==2024.1
